+++ R/utils-options.R
@@ -471,5 +471,6 @@ update_package_options <- function(debug = FALSE) {
   update_package_option("future.future.gc", mode = "character", debug = debug)
   update_package_option("future.future.local", mode = "character", debug = debug)
   update_package_option("future.plan.earlySignal", mode = "character", debug = debug)
+  update_package_option("future.resolved.run", mode = "character", debug = debug)
   update_package_option("future.Future.envir.keep", mode = "logical", debug = debug)
 }